Move Address class to address individual package
